refactor(router): move Router module out of vehicle/ to tesla/router.py
The router module exports the entity-agnostic Router base plus the
VehicleRouter and EnergySiteRouter peers, so it is not vehicle-specific.
Relocate it next to energysite.py and fix every import site.

- git mv tesla/vehicle/router.py -> tesla/router.py (logic unchanged)
- git mv tests/test_vehicle_router.py -> tests/test_router.py (now also
  covers energy-site and N-way routing); broaden its module docstring
- tesla/__init__.py: import the three names from tesla.router; __all__
  still exports all three (top-level public API preserved)
- tesla/vehicle/__init__.py: drop the router re-exports
- README.md / AGENTS.md: point docs at tesla.router

Clean break, no back-compat shim: these classes are unreleased (landed in
#37/#38, after v1.4.7), so no published artifact exposed the old
tesla.vehicle.router path.
