+++ tesla_fleet_api/tesla/__init__.py
@@ -6,6 +6,7 @@ from tesla_fleet_api.tesla.oauth import TeslaFleetOAuth
 from tesla_fleet_api.tesla.charging import Charging
 from tesla_fleet_api.tesla.energysite import EnergySites, EnergySite
 from tesla_fleet_api.tesla.partner import Partner
+from tesla_fleet_api.tesla.router import Router, VehicleRouter, EnergySiteRouter
 from tesla_fleet_api.tesla.user import User
 from tesla_fleet_api.tesla.vehicle import (
     Vehicles,
@@ -14,9 +15,6 @@ from tesla_fleet_api.tesla.vehicle import (
     VehicleSigned,
     VehicleBluetooth,
     Vehicle,
-    Router,
-    VehicleRouter,
-    EnergySiteRouter,
 )
 
 __all__ = [
