+++ tesla_fleet_api/tesla/vehicle/__init__.py
@@ -4,7 +4,6 @@ from tesla_fleet_api.tesla.vehicle.vehicles import Vehicles, VehiclesBluetooth
 from tesla_fleet_api.tesla.vehicle.fleet import VehicleFleet
 from tesla_fleet_api.tesla.vehicle.bluetooth import VehicleBluetooth
 from tesla_fleet_api.tesla.vehicle.signed import VehicleSigned
-from tesla_fleet_api.tesla.vehicle.router import Router, VehicleRouter, EnergySiteRouter
 from tesla_fleet_api.tesla.vehicle.vehicle import Vehicle
 
 __all__ = [
@@ -14,7 +13,4 @@ __all__ = [
     "VehicleFleet",
     "VehicleBluetooth",
     "VehicleSigned",
-    "Router",
-    "VehicleRouter",
-    "EnergySiteRouter",
 ]
